Move the MockShellOut class into its own file.
