+++ test/cassandra/compaction_check_test.rb
@@ -1,19 +1,5 @@
 require 'test_helper'
-
-class MockShellOut
-  attr_reader :stdout
-  attr_accessor :valid_exit_codes
-
-  def initialize(stdout)
-    @stdout = stdout
-  end
-
-  def run_command
-  end
-
-  def error!
-  end
-end
+require 'mock_shell_out'
 
 describe Cassandra::Utils::Stats::Compaction do
   before do
