+++ test/mock_shell_out.rb
@@ -1,0 +1,14 @@
+class MockShellOut
+  attr_reader :stdout
+  attr_accessor :valid_exit_codes
+
+  def initialize(stdout)
+    @stdout = stdout
+  end
+
+  def run_command
+  end
+
+  def error!
+  end
+end
